Rename sidecar doctor surfaces to diagnostics
