+++ lib/host_config/host_config.mli
@@ -71,7 +71,7 @@ type t =
   ; base_path_raw : string option
         (** [MASC_BASE_PATH] value as read from the environment, with
             whitespace trimmed but no path normalisation.  Used by
-            routes / dashboard / config_doctor inputs that surface
+            routes / dashboard / config diagnostics inputs that surface
             the operator's literal input.  RFC-0085 PR-9 replaces
             [Env_config_core.base_path_raw_opt]. *)
   ; config_dir : string option
