+++ sidecars/imessage-bot/run.sh
@@ -12,9 +12,9 @@
 #   ./run.sh              start the bot (foreground, tees to today's log)
 #   ./run.sh tail         tail -F today's log file
 #   ./run.sh status       pretty-print the current status.json
-#   ./run.sh doctor       run health checks without starting the bot
-#   ./run.sh doctor --json   machine-readable check output
-#   ./run.sh doctor --fix    run available auto-fixes then recheck
+#   ./run.sh diagnostics       run health checks without starting the bot
+#   ./run.sh diagnostics --json   machine-readable check output
+#   ./run.sh diagnostics --fix    run available auto-fixes then recheck
 
 set -euo pipefail
 
@@ -73,10 +73,10 @@ case "$cmd" in
       cat "$STATUS_FILE"
     fi
     ;;
-  doctor)
+  diagnostics)
     cd "$(script_dir)"
     shift || true
-    python -m src doctor "$@"
+    python -m src diagnostics "$@"
     ;;
   stop)
     # Match by absolute script_dir path so we never hit another sidecar.
@@ -88,7 +88,7 @@ case "$cmd" in
     fi
     ;;
   *)
-    echo "Usage: $0 [start|stop|tail|status|doctor]" >&2
+    echo "Usage: $0 [start|stop|tail|status|diagnostics]" >&2
     exit 2
     ;;
 esac
