+++ sidecars/imessage-bot/src/__main__.py
@@ -1,11 +1,11 @@
 """Entry point: python -m src.
 
-기본은 봇 기동. ``doctor`` subcommand 로 건강 점검만 실행.
+기본은 봇 기동. ``diagnostics`` subcommand 로 건강 점검만 실행.
 
     python -m src                # 봇 기동
-    python -m src doctor         # 점검
-    python -m src doctor --json  # 점검 결과 JSON
-    python -m src doctor --fix   # 가능한 auto-fix 후 재점검
+    python -m src diagnostics         # 점검
+    python -m src diagnostics --json  # 점검 결과 JSON
+    python -m src diagnostics --fix   # 가능한 auto-fix 후 재점검
 """
 from __future__ import annotations
 
@@ -26,26 +26,26 @@ from gate_shared import (  # noqa: E402
 )
 
 
-def _run_doctor(argv: list[str]) -> int:
-    # Lazy import — doctor must still run when runtime deps are missing.
-    from .doctor import run_doctor  # noqa: PLC0415
+def _run_diagnostics(argv: list[str]) -> int:
+    # Lazy import — diagnostics must still run when runtime deps are missing.
+    from .diagnostics import run_diagnostics  # noqa: PLC0415
 
     as_json = "--json" in argv
     auto_fix = "--fix" in argv
 
     async def run() -> int:
-        doctor = await run_doctor()
-        checks = await doctor.run()
+        diagnostics = await run_diagnostics()
+        checks = await diagnostics.run()
         outcomes: list[FixOutcome] = []
         if auto_fix:
-            checks, outcomes = await doctor.run_auto_fixes(checks)
+            checks, outcomes = await diagnostics.run_auto_fixes(checks)
         if as_json:
-            sys.stdout.write(render_json(doctor.title, checks))
+            sys.stdout.write(render_json(diagnostics.title, checks))
         else:
             fix_block = render_fix_outcomes(outcomes)
             if fix_block:
                 sys.stdout.write(fix_block + "\n")
-            sys.stdout.write(render_pretty(doctor.title, checks))
+            sys.stdout.write(render_pretty(diagnostics.title, checks))
         sys.stdout.write("\n")
         return exit_code_for(checks)
 
@@ -53,8 +53,8 @@ def _run_doctor(argv: list[str]) -> int:
 
 
 if __name__ == "__main__":
-    if len(sys.argv) > 1 and sys.argv[1] == "doctor":
-        raise SystemExit(_run_doctor(sys.argv[2:]))
+    if len(sys.argv) > 1 and sys.argv[1] == "diagnostics":
+        raise SystemExit(_run_diagnostics(sys.argv[2:]))
     from .bot import main  # noqa: PLC0415
 
     asyncio.run(main())
